refactor: adjust api route based on review feedback
- adjust api endpoint logic
- refactor fetch logic into own reusable component
